% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_layer_hierarchy.R
\name{build_layer_hierarchy}
\alias{build_layer_hierarchy}
\title{Work out the layer hierarchy in the Distance database}
\usage{
build_layer_hierarchy(data_file)
}
\arguments{
\item{data_file}{a data file to load the database from}
}
\description{
Use the DataLayers table to work out the hierarchy of the tables and layers in the database held by Distance.
}
\author{
David L Miller
}

