% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_project.R
\name{convert_project}
\alias{convert_project}
\title{Convert a Distance for Windows project}
\usage{
convert_project(project)
}
\arguments{
\item{project}{a path to a project}
}
\description{
Summary here
}
\section{Details}{

At the moment only line or point transects are supported.
}

