% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdst-package.R
\name{converted_distance_analyses}
\alias{converted_distance_analyses}
\alias{converted_distance_analysis}
\title{Converted analysis objects}
\description{
Once \code{\link{convert_project}} has been run on a project, two types of object are created: first an object of class \code{converted_distance_analyses}, which is just a list of \code{converted_distance_analysis} objects.
}
\details{
\code{converted_distance_analysis} contain all the information necessary to run a Distance for Windows model in R. Each object has the following elements:

\itemize{
  \item{\code{call}}{ string with the call to \code{\link{ddf}} to build and run the model}
  \item{\code{aic.select}}{ maximum number of terms to select by AIC if AIC term selection has been enabled (for key plus adjustment terms models only)}
  \item{\code{status}}{ what the status of this model was in Distance for Windows (see "Status" below)}
  \item{\code{env}}{ an \code{environment} that contains data needed to run the model (\code{data} containing entire dataset in flatfile form, \code{obs.table} containing the observation table, \code{sample.table} is the sample table, \code{reg.table} is the region table and \code{units} is a matrix describing conversion factor of distance measures (effort and detection distance) to areal measurements (for density))}
  \item{\code{filter}}{ string used to subset the data to get the same filter as in Distance for Windows}
  \item{\code{group_size}}{ describes how size bias adjustment is conducted, and the level of hierarchy at which E(s) is computed}
  \item{\code{detection_by}}{ level of design hierarchy at which detection function is computed (e.g. pooled across strata)}
  \item{\code{gof_intervals}}{ if binning is done for GOF testing, cutpoints are provided here}
  \item{\code{estimation}}{ what sort of weighted average is used to compute region-level density estimate}
  \item{\code{name}}{ the name for this analysis, as used in Distance for Windows}
  \item{\code{ID}}{ the ID number for this analysis, as used in Distance for Windows}
}
}
\section{Status}{

The \code{status} code is taken from Distance for Windows to indicate whether the analysis has been run yet and what the outcome was. Status codes are as follows:

\itemize{
  \item{\code{0}}{ analyses has not been run in Distance for Windows yet}
  \item{\code{1}}{ analysis ran without errors or warnings}
  \item{\code{2}}{ analysis ran with warnings}
  \item{\code{3}}{ analysis ran with errors}
}

Note that an analysis that runs with error in Distance for Windows may run fine in R and an analysis that runs fine in Distance for Windows may not work in R. In the latter case, please consider submitting this a a bug to \url{github.com/distancedevelopment/distance-bugs}.
}

