% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdst-package.R
\name{converted_distance_data}
\alias{converted_distance_data}
\title{Converted distance data}
\description{
If \code{\link{convert_project}} has been run on a project, but there are no analyses present in the project, then a list of the data will be returned. The list has one element for each data filter which was present in the project.
Each element of the list has the following tables in it:
}
