% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get.R
\name{db_get}
\alias{db_get}
\title{Get data from the Distance project database}
\usage{
db_get(file, table = NULL)
}
\arguments{
\item{file}{the path to the database file to access}

\item{table}{the table to extract (if \code{NULL} all tables are extracted, if \code{TRUE} a list of tables names are extracted)}
}
\value{
a \code{data.frame} with the contents of a database table
}
\description{
This function is a wrapper around either calls to \code{\link{RODBC}} (on Windows) or \code{\link{mdb.get}} (on Unix-a-like systems). Given a database file name it will return either the contents of the table (as \code{data.frame}), if \code{table=NULL} then it will return all tables and if \code{table=TRUE} then it will return a character vector of table names.
}
\note{
Currently not implemented on Windows systems.
}
\author{
David L Miller
}
