% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter the data}
\usage{
filter_data(data, data_filter)
}
\arguments{
\item{data}{the data to be filtered}

\item{data_filter}{a data filter to be parsed (output from \code{\link{parse_definition.data_filter}})}
}
\value{
a list with two elements, the data and the filter string
}
\description{
Take the "data filters" applied by Distance for Windows to the data and use them to subset the data.
}
\author{
David L Miller
}
