% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Extract data from a Distance database}
\usage{
get_data(data_file)
}
\arguments{
\item{data_file}{the path to a \code{DistData.mdb} file.}
}
\description{
Extracts the relevant (existing) tables from the Distance database.
}
\details{
Currently, this only returns the observation table, eventually it will return all tables.
}
\author{
David L Miller
}

