% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_definitions.R
\name{get_definitions}
\alias{get_definitions}
\title{Extract definition information from tables}
\usage{
get_definitions(file, table)
}
\arguments{
\item{file}{the name of the \code{.mdb} file}

\item{table}{which table to extract the "Description" column from}
}
\value{
a \code{list} of definitions, each element of which is a character vector.
}
\description{
Takes the "Definition" column from a table, converts it to \code{character} and puts each row in a list with names as the corresponding ID of that row.
}
\author{
David L Miller
}

