% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Extract saved statistics for analyses}
\usage{
get_stats(project_file)
}
\arguments{
\item{project_file}{path to project file}
}
\description{
At the moment only extracts the AIC and likelihood
}
\author{
David L Miller
}

