% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unit_conversion.R
\name{get_unit_conversion}
\alias{get_unit_conversion}
\title{Get the unit conversions for the data}
\usage{
get_unit_conversion(data_file)
}
\arguments{
\item{data_file}{Distance for Windows project data file}
}
\value{
a \code{data.frame} with columns \code{Variable}, \code{Units} and \code{Conversion}, giving the variable name, the units it is measured in and the conversation factor to SI units.
}
\description{
Obtain a list of conversion to SI units from the units that the measurements in a Distance for Windows project are in.
}
\author{
David L Miller
}

