% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_size_est.R
\name{group_size_est}
\alias{group_size_est}
\title{Estimate group size}
\usage{
group_size_est(data, model_definition, model)
}
\arguments{
\item{data}{the data for this analysis}

\item{model_definition}{a model definition}

\item{model}{a fitted model}
}
\value{
a single the estimated cluster size, or \code{NULL} if there were no instructions on how to estimate group/cluster size
}
\description{
Distance for Windows includes a few different methods for accounting for group (or cluster) size (at the adundance/density estimation stage). These include using the mean group size for all observations or using a regression of size against distance or log size against distance.
}
\author{
David L Miller
}

