% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_analysis.R
\name{make_analysis}
\alias{make_analysis}
\title{Make an analysis}
\usage{
make_analysis(this_analysis, model_definitions, data_filters, data, transect)
}
\description{
This function calls \code{\link{make_model}} to create the call to \code{\link{ddf}} it also creates an environment with the data necessary to perform the call.
}
\author{
David L Miller
}

