% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_control.R
\name{make_control}
\alias{make_control}
\title{Make the control element of a call to \code{ddf}}
\usage{
make_control(md)
}
\arguments{
\item{md}{model definition data to parse}
}
\value{
character string describing the \code{control} \code{list}
}
\description{
Build the \code{control} options for a \code{ddf} call.
}
