% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_model.R
\name{make_model}
\alias{make_model}
\title{Build a distance sampling analysis}
\usage{
make_model(this_analysis, model_definitions, data_filters, transect, data)
}
\arguments{
\item{this_analysis}{an analysis from Distance}

\item{model_definitions}{a list of model definitions}

\item{data_filters}{a list of data filters}

\item{transect}{the transect type}

\item{data}{the data}
}
\value{
a character string specifying a call to \code{ddf}
}
\description{
Reproduce the corresponding call to \code{ddf} to reproduce an analysis from Distance for Windows.
}
\author{
David L Miller
}

