% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mrmodel.R
\name{make_mrmodel}
\alias{make_mrmodel}
\title{Build a mrmodel call}
\usage{
make_mrmodel(md)
}
\arguments{
\item{md}{a model definition}
}
\value{
a character string starting with "\code{mrmodel=}" or \code{NULL} if there is no \code{mrmodel} component in this model.
}
\description{
From a model definition build the \code{mrmodel} part of the model.
}
\author{
David L Miller
}
