% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results.R
\name{merge_results}
\alias{merge_results}
\title{Merge results from stratified analyses}
\usage{
merge_results(models, analysis)
}
\arguments{
\item{models}{a \code{list} of model (\code{ddf}/\code{ds} objects)}

\item{analysis}{an analysis specification (to inform us on the stratification to be used.}
}
\value{
a list including the "combined" model, summary and density/abundance estimates (\code{dht} output). Note that these are almost definitely not valid objects for their respective classes, they are only to be used to test statistics.
}
\description{
In Distance for Windows, one can choose to estimate the detection function by stratum. In this case more than one detection function is returned when \code{\link{run_analysis}} is used to run the analysis. In order to test the statistics stored in the Distance for Windows project, one must first combine the resulting models (and their corresponding abundance and density estimates). This function performs these operations.
}
\author{
David L Miller
}

