% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_description.R
\name{model_description}
\alias{model_description}
\title{Make a user-readable model description string}
\usage{
model_description(model)
}
\arguments{
\item{model}{a fitted model}
}
\value{
a string describing the model
}
\description{
This takes a fitted \code{mrds} model object and returns a string that describes the detection function fitted and the fitted model's AIC.
}
\author{
David L Miller
}
