% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Model selection for key plus adjustment models}
\usage{
model_selection(analysis, debug = FALSE)
}
\arguments{
\item{analysis}{a converted analysis}

\item{debug}{display the call and name of the model before it is run, print AIC selection details}
}
\value{
fitted \code{\link{ddf}} object
}
\description{
Run model selection for a given analysis. The returned object is exactly as if the model has been run using \code{\link{ddf}}, so anything that can normally be done with a \code{ddf} object can be done with the return.
}
\details{
Model selection is performed via AIC.
}
\author{
David L Miller
}

