% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_definition.model.R
\name{parse_definition.model}
\alias{parse_definition.model}
\title{Parse a Definition}
\usage{
parse_definition.model(df)
}
\arguments{
\item{df}{a definition (vector of character strings)}
}
\value{
a \code{list} of \code{list}s
}
\description{
Given data from a "Definition", pre-processed by \code{\link{get_definitions}}, extract the useful information from it.
}
\section{Details}{

See the "MCDS Command Language" section of the Distance manual for more information.

Note that this function should be called for a single definition, usually using \code{\link{lapply}}.
}
\author{
David L Miller
}

