% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.converted_distance_analyses.R
\name{print.converted_distance_analyses}
\alias{print.converted_distance_analyses}
\title{Print a converted distance analysis}
\usage{
\method{print}{converted_distance_analyses}(x, ...)
}
\description{
Prints a table of the analyses that have been converted.
}

