% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.converted_distance_analysis.R
\name{print.converted_distance_analysis}
\alias{print.converted_distance_analysis}
\title{Print a converted distance analysis}
\usage{
\method{print}{converted_distance_analysis}(x, ...)
}
\description{
Prints details of an analyses that has been converted.
}

