% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdst-package.R
\docType{package}
\name{readdst-package}
\alias{readdst}
\alias{readdst-package}
\title{Convert Distance for Windows analyses to R code}
\description{
This package read data and model definitions from a Distance for Windows project (\code{.dst} and \code{.dat} files) and converts models to run in the R package \code{\link{mrds}}.
}

