% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run a converted Distance analysis}
\usage{
run_analysis(analysis, debug = FALSE)
}
\arguments{
\item{analysis}{a converted analysis}

\item{debug}{display the call and name of the model before it is run}
}
\value{
fitted \code{\link{ddf}} object
}
\description{
Take a single converted analysis and run the model contained therein.
}
\author{
David L Miller
}

