% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_table.R
\name{stats_table}
\alias{stats_table}
\title{Generate table of possible statistics to test}
\usage{
stats_table()
}
\value{
a \code{data.frame} with statistics Distance for Windows collects that have equivalents in \code{mrds}. The \code{data.frame} has three columns: \code{Code}, the numeric code for the statistic (as used in the Distance for Windows database); \code{Name}, the short name for this statistic; \code{MRDS}, the operation required to obtain the equivalent statistic in \code{mrds}; \code{Description}, a short description of the statistic.
}
\description{
To use \code{\link{get_stats}} we need a set of statistics to test. We also require their codes (to look up in the Distance for Windows database) and their equivalent values in \code{mrds} (or how to calculate those values). This function provides such a table.
}
\section{Details}{

Data for this table (numeric code and descriptions) is from the \code{DistIni.mdb} which is shipped with Distance for Windows. See also \url{https://github.com/distancedevelopment/readdst/wiki/distance-results-codes}.
}

\section{Additional notes}{

Note that the Cramer-von Mises p-value as recorded in Distance for Windows is only recorded to the nearest 0.1.
}
\author{
David L Miller
}

