% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_stats.R
\name{test_stats}
\alias{test_stats}
\title{Test to see if Distance and R get the same result}
\usage{
test_stats(analysis, project)
}
\arguments{
\item{analysis}{a converted (but not run) analysis}

\item{project}{a project path}
}
\description{
Test to see if Distance and R get the same result
}

