% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unflatfile.R
\name{unflatfile}
\alias{unflatfile}
\title{Take a flatfile data.frame and make dht-compatible data.frames}
\usage{
unflatfile(data)
}
\value{
list of the four \code{data.frame}s described in "Details".
}
\description{
Given distance sampling survey data in flatfile format, convert it to the four tables required by \code{\link{dht}}.
}
\note{
Based on \code{checkdata} from package \code{Distance}.
}
\section{Details}{


\itemize{
\item{region.table } \code{data.frame} with two columns: \code{Region.Label}, label for the region; \code{Area}, area of the region. \code{region.table} has one row for each stratum. If there is no stratification then \code{region.table} has one entry with \code{Area} corresponding to the total survey area.
\item{sample.table } \code{data.frame} mapping the regions to the samples (i.e. transects). There are three columns: \code{Sample.Label}, label for the sample; \code{Region.Label}, label for the region that the sample belongs to.; \code{Effort}, the effort expended in that sample (e.g. transect length).
\item{obs.table } \code{data.frame} mapping the individual observations (objects) to regions and samples. There should be three columns: \code{object}, the observation ID; \code{Region.Label}, label for the region that the sample belongs to; \code{Sample.Label}, label for the sample.
\item{data } a \code{data.frame} containing at least a column called \code{distance}. NOTE! If there is a column called \code{size} in the data then it will be interpreted as group/cluster size.
}
}
\author{
David L. Miller
}

