% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_project.R
\name{convert_project}
\alias{convert_project}
\title{Convert a Distance for Windows project}
\usage{
convert_project(project)
}
\arguments{
\item{project}{a path to a project}
}
\value{
an object of class \code{converted_distance_analyses}
}
\description{
Take each analysis in a Distance for Windows project and convert the model definition to an \code{mrds} model. Note that at the moment only CDS/MCDS/MRDS analyses are supported.
}
\details{
Note that model names are as they are in Distance for Windows (so if you have nonsensical names in Distance for Windows they will be the same in R).
}
\section{Details}{

At the moment only line or point transects are supported.
}

