% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Extract data from a Distance database}
\usage{
get_data(data_file)
}
\arguments{
\item{data_file}{the path to a \code{DistData.mdb} file.}
}
\value{
a "flatfile" compatible \code{data.frame} containing all of the information necessary to make a stratified abundance/density estimate.
}
\description{
Extracts the relevant tables from the Distance for Windows database to build data that can be used with \code{mrds} or \code{Distance}.
}
\author{
David L Miller
}
