% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_dsmodel.R
\name{make_dsmodel}
\alias{make_dsmodel}
\title{Build a dsmodel call}
\usage{
make_dsmodel(md)
}
\arguments{
\item{md}{a model definition}
}
\value{
a character string starting with \code{dsmodel=}
}
\description{
From a model definition build the \code{dsmodel} part of the model.
}
\author{
David L Miller
}

