% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_meta.data.R
\name{make_meta.data}
\alias{make_meta.data}
\title{Build the model meta.data}
\usage{
make_meta.data(df, transect)
}
\arguments{
\item{df}{a data filter object}

\item{transect}{type of transect}
}
\value{
a character string starting with \code{meta.data=}
}
\description{
From a model definition build the \code{dsmodel} part of the model.
}
\author{
David L Miller
}

