% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_definition.model.R
\name{parse_definition.model}
\alias{parse_definition.model}
\title{Parse a Definition}
\usage{
parse_definition.model(df)
}
\arguments{
\item{df}{a definition}
}
\value{
named vector of defintions
}
\description{
Given data from a "Definition", pre-processed by \code{\link{get_definitions}}, extract the useful information from it.
}
\section{Details}{

A definition consists either of a key=value pair or a name then key=value pairs separated by \code{\\} and terminated with \code{;}.

Note that this function should be called for a single definition, usually using \code{\link{lapply}}.
}
\author{
David L Miller
}

