% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.converted_distance_analyses.R
\name{print.converted_distance_analyses}
\alias{print.converted_distance_analyses}
\title{Converted distance analyses table}
\usage{
\method{print}{converted_distance_analyses}(x, ...)
}
\arguments{
\item{x}{converted distance analyses}

\item{...}{unused additional args for S3 compatibility}
}
\description{
Prints a table of the analyses that have been converted and their status from Distance for Windows.
}
