% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stats.R
\name{print.distance_stats_table}
\alias{print.distance_stats_table}
\title{Print tested statistics}
\usage{
\method{print}{distance_stats_table}(x)
}
\arguments{
\item{x}{the result of a call to \code{\link{test_stats}}}
}
\value{
just prints the results
}
\description{
This is simply a \code{print} method to nicely ouput the results of \code{\link{test_stats}}.
}
\author{
David L Miller
}

