% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_covar_names.R
\name{set_covar_names}
\alias{set_covar_names}
\title{Set column names in data to be as in formulae}
\usage{
set_covar_names(data, covnames)
}
\arguments{
\item{data}{a \code{data.frame} of the data to be modelled}

\item{covnames}{the covariates that are factors}
}
\description{
Set column names in data to be as in formulae
}

