% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stats.R
\name{test_stats}
\alias{test_stats}
\title{Test to see if Distance for Windows and R get the same results}
\usage{
test_stats(analysis, statuses = 1)
}
\arguments{
\item{analysis}{a converted (but not run) analysis}

\item{statuses}{for which statuses should tests be run? Defaults to \code{1} but can be given values like \code{1:3} ("Run", "Run with warnings", "Run with errors", respectively).}
}
\description{
Tests the results stored in the Distance for Windows project file against those generated from running the same analysis in R.
}
\details{
Currently only tests AIC and log-likelihood values.
}

